Release 3.1.0

- chore: Update copyright date in the SPDX headers
- fix: Fix typo in an error message specifiy => specify
- fix: Warning: comparison of integers of different signs
- fix: Fix warning: variable 'ret' is used uninitialized
- fix: Fix tpm2.h include guard, and use #pragma once for all headers
- feat: Load i2c-dev through /etc/modules-load.d
- chore: Use GNUInstallDirs cmake-module to define install location
- chore: Remove C++ dependency from project
- chore: Use CMake project version as PISERIAL_VERSION
- chore: Make repository REUSE compliant
- ci: Add pipeline for package build

+++ src/atecc508a.c
@@ -1,8 +1,6 @@
-/* SPDX-License-Identifier: GPL-2.0-or-later */
-
-/*
- * Copyright: 2021-2023 KUNBUS GmbH
- */
+// SPDX-FileCopyrightText: 2021-2023 KUNBUS GmbH <support@kunbus.com>
+//
+// SPDX-License-Identifier: GPL-2.0-or-later
 
 /*
  * Microchip Technology Inc. ATECC508A
@@ -193,7 +191,7 @@ static int atecc508a_read(const struct atecc508a *const priv,
 			retry++;
 			if (retry >= MAX_RETRY)
 				return -1;
-		} else if (ret != sizeof(struct atecc508a_recv_data) - num_read) {
+		} else if ((size_t)ret != sizeof(struct atecc508a_recv_data) - num_read) {
 			num_read += ret;
 			continue;
 		} else {
